+++ src/modules/edc-demo/components/asset-detail-dialog/asset-detail-dialog-field-builder.ts
@@ -4,11 +4,9 @@ import {Asset} from '../../models/asset';
 import {validUrlPattern} from '../../validators/url-validator';
 import {AssetDetailDialogField} from './asset-detail-dialog-field';
 
-
 @Injectable()
 export class AssetDetailDialogFieldBuilder {
-  constructor(private activeFeatureSet: ActiveFeatureSet) {
-  }
+  constructor(private activeFeatureSet: ActiveFeatureSet) {}
 
   buildFields(asset: Asset): AssetDetailDialogField[] {
     const tryValue = (s?: string | null) => s || '-';
@@ -93,7 +91,7 @@ export class AssetDetailDialogFieldBuilder {
         icon: 'account_circle',
         label: 'Organization',
         text: tryValue(asset.originatorOrganization),
-      }
+      },
     );
 
     return fields;