+++ src/modules/edc-demo/services/contract-negotiation.service.ts
@@ -50,9 +50,7 @@ export class ContractNegotiationService {
     }
   }
 
-  updateNegotiationList() {
-
-  }
+  updateNegotiationList() {}
 
   isBusy(contractOffer: ContractOffer) {
     return (